+++ src/cmos/time.rs
@@ -1,10 +1,10 @@
 use core::fmt;
 
-const MINUTE: u32 = 60;
-const HOUR: u32 = MINUTE * 60;
-const DAY: u32 = HOUR * 24;
-const YEAR: u32 = DAY * 365;
-static MONTH: [u32; 12] = [
+const MINUTE: u64 = 60;
+const HOUR: u64 = MINUTE * 60;
+const DAY: u64 = HOUR * 24;
+const YEAR: u64 = DAY * 365;
+static MONTH: [u64; 12] = [
     0,
     DAY * (31),
     DAY * (31 + 29),
@@ -29,9 +29,9 @@ pub struct DateTime {
 }
 
 impl DateTime {
-    pub fn timestamp(&self) -> u32 {
-        let mut timestamp: u32 = 0;
-        let mut myear: u32 = self.year as u32;
+    pub fn timestamp(&self) -> u64 {
+        let mut timestamp: u64 = 0;
+        let mut myear: u64 = self.year as u64;
 
         if self.year < 70 {
             // mOS will not work proprely after 2069 :-)
@@ -40,10 +40,10 @@ impl DateTime {
             myear -= 70;
         }
 
-        timestamp += self.seconds as u32;
-        timestamp += (self.minutes as u32) * MINUTE;
-        timestamp += (self.hour as u32) * HOUR;
-        timestamp += ((self.day as u32) - 1) * DAY; // -1 because day is in progress :-)
+        timestamp += self.seconds as u64;
+        timestamp += (self.minutes as u64) * MINUTE;
+        timestamp += (self.hour as u64) * HOUR;
+        timestamp += ((self.day as u64) - 1) * DAY; // -1 because day is in progress :-)
         timestamp += MONTH[(self.month - 1) as usize];
         timestamp += myear * YEAR + DAY * ((myear + 1) / 4);
 
@@ -80,7 +80,7 @@ impl fmt::Display for DateTime {
     }
 }
 
-fn is_leap_year(year: u32) -> bool {
+fn is_leap_year(year: u64) -> bool {
     return (year % 4 == 0) && (year % 100 != 0 || year % 400 == 0);
 }
 
@@ -155,7 +155,7 @@ mod test {
 
     #[test]
     fn test_is_leap_year() {
-        let leap_years: [u32; 37] = [
+        let leap_years: [u64; 37] = [
             1904, 1908, 1912, 1916, 1920, 1924, 1928, 1932, 1936, 1940, 1944, 1948, 1952, 1956,
             1960, 1964, 1968, 1972, 1976, 1980, 1984, 1988, 1992, 1996, 2000, 2004, 2008, 2012,
             2016, 2020, 2024, 2028, 2032, 2036, 2040, 2044, 2048,
@@ -167,7 +167,7 @@ mod test {
 
     #[test]
     fn test_is_not_leap_year() {
-        let not_leap_years: [u32; 26] = [
+        let not_leap_years: [u64; 26] = [
             1905, 1909, 1911, 1917, 1921, 1922, 1923, 1925, 1926, 1927, 1929, 1930, 1931, 2007,
             2009, 2010, 2011, 2013, 2014, 2015, 2017, 2018, 2019, 2021, 2022, 2023,
         ];
